Move function metering tests to interpreter module
